Take files to watch from stdin instead of -d
Running tests modifies .pyc files, so after
modifying a python file, the tests would run twice:
once for the .py file, and again for the .pyc

So, some sort of exclude functionality was necessary.

Instead of reimplementing what is essentially 'find',
now watchdir will read from STDIN the list of files
it needs to watch.